backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
bcrypt==4.1.2
PyJWT==2.8.0
pydantic[email]==2.5.0
python-multipart==0.0.6
pillow==10.0.0
numpy==1.24.3
insightface==0.7.3
onnxruntime==1.16.3
opencv-python-headless==4.8.0.76
scikit-learn==1.3.0
scikit-image==0.21.0
matplotlib==3.7.2
scipy==1.11.1 
awscli==1.32.106
